+++ src/Sonarr.Api.V5/EpisodeFiles/EpisodeFileController.cs
@@ -127,12 +127,26 @@ public class EpisodeFileController : RestControllerWithSignalR<EpisodeFileResour
     [Consumes("application/json")]
     public object DeleteEpisodeFiles([FromBody] EpisodeFileListResource resource)
     {
-        var episodeFiles = _mediaFileService.GetFiles(resource.EpisodeFileIds);
-        var series = _seriesService.GetSeries(episodeFiles.First().SeriesId);
-
-        foreach (var episodeFile in episodeFiles)
+        if (resource.EpisodeFileIds.Count == 0)
         {
-            _mediaFileDeletionService.DeleteEpisodeFile(series, episodeFile);
+            return NoContent();
+        }
+
+        var episodeFiles = _mediaFileService.GetFiles(resource.EpisodeFileIds);
+
+        if (episodeFiles.Count == 0)
+        {
+            return NoContent();
+        }
+
+        foreach (var seriesFiles in episodeFiles.GroupBy(e => e.SeriesId))
+        {
+            var series = _seriesService.GetSeries(seriesFiles.Key);
+
+            foreach (var episodeFile in seriesFiles)
+            {
+                _mediaFileDeletionService.DeleteEpisodeFile(series, episodeFile);
+            }
         }
 
         return NoContent();
